pyproject.toml
[project]
name = "toronto-streetview-count"
version = "0.1.0"
description = "Count Google Street View panoramas in Toronto using road network sampling"
readme = "README.md"
authors = [
    { name = "Alex Olson", email = "alex.olson@utoronto.ca" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "httpx>=0.28.1",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
    "pyogrio>=0.6.0",
            "aiosqlite>=0.19.0",
        "pydantic>=2.0.0",
        "overpy>=0.7.0",
        "pyarrow>=12.0.0",
        "toronto-open-data>=0.1.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[project.scripts]
toronto-streetview-count = "toronto_streetview_count:main"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
